Move out different key conditions to separate files
